tools/python/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==20.3.0
autoflake==1.4
Babel==2.9.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
commonmark==0.9.1
docutils==0.17
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isodate==0.6.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
livereload==2.6.3
lxml==4.6.3
m2r==0.2.1
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pluggy==0.13.1
pudb==2020.1
py==1.10.0
pyflakes==2.3.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-pudb==0.7.0
pytz==2021.1
PyYAML==5.4.1
recommonmark==0.7.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-openapi==0.7.0
sphinxcontrib-programoutput==0.17
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
unify==0.5
untokenize==0.1.1
urllib3==1.26.4
urwid==2.1.2
yamllint==1.26.1
